# Fix JsonSerDe parsing performance